small notebook refactor
