+++ src/components/tiptap-templates/simple/toolbar/MobileToolbar.tsx
@@ -1,0 +1,39 @@
+import {
+  ToolbarGroup,
+  ToolbarSeparator,
+} from "@/components/tiptap-ui-primitive/toolbar";
+import { Button } from "@/components/tiptap-ui-primitive/button";
+import { ArrowLeftIcon } from "@/components/tiptap-icons/arrow-left-icon";
+import { HighlighterIcon } from "@/components/tiptap-icons/highlighter-icon";
+import { LinkIcon } from "@/components/tiptap-icons/link-icon";
+import { ColorHighlightPopoverContent } from "@/components/tiptap-ui/color-highlight-popover";
+import { LinkContent } from "@/components/tiptap-ui/link-popover";
+
+export const MobileToolbarContent = ({
+  type,
+  onBack,
+}: {
+  type: "highlighter" | "link";
+  onBack: () => void;
+}) => (
+  <>
+    <ToolbarGroup>
+      <Button data-style="ghost" onClick={onBack}>
+        <ArrowLeftIcon className="tiptap-button-icon" />
+        {type === "highlighter" ? (
+          <HighlighterIcon className="tiptap-button-icon" />
+        ) : (
+          <LinkIcon className="tiptap-button-icon" />
+        )}
+      </Button>
+    </ToolbarGroup>
+
+    <ToolbarSeparator />
+
+    {type === "highlighter" ? (
+      <ColorHighlightPopoverContent />
+    ) : (
+      <LinkContent />
+    )}
+  </>
+);
